chore: rename to custom-elements-json
